# src/backend/api-spec.yaml
openapi: 3.0.0
info:
  title: Hospital Network Management System API
  description: |
    API specification for the Hospital Network Management System,
    enabling secure interoperability between hospital systems.
  version: 1.0.0
servers:
  - url: https://api.hospital-network.org/v1
    description: Production API server
  - url: https://staging-api.hospital-network.org/v1
    description: Staging API server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Hospital:
      type: object
      required:
        - name
        - address
        - contactEmail
        - contactPhone
      properties:
        hospitalId:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        address:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
  /hospitals:
    get:
      summary: List all hospitals in the network
      operationId: listHospitals
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of hospitals
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Hospital'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Register a new hospital
      operationId: createHospital
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hospital'
      responses:
        '201':
          description: Hospital created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hospital'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /hospitals/{hospitalId}:
    get:
      summary: Get hospital details
      operationId: getHospital
      parameters:
        - name: hospitalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Hospital details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hospital'
        '404':
          description: Hospital not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update hospital details
      operationId: updateHospital
      parameters:
        - name: hospitalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hospital'
      responses:
        '200':
          description: Hospital updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hospital'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Hospital not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...